+++ RELEASE_NOTES.md
@@ -1,4 +1,5 @@
 ### 2.1 - Unreleased
+* Regex partial active pattern
 
 ### 2.0 - 2020/08/09
 * (Breaking change) Update.run signature changed